fix #3497: Split experiment test factories
* Move Nimbus factories into their own submodule
