+++ app/experimenter/experiments/tests/factories/__init__.py
@@ -1,0 +1,14 @@
+from experimenter.experiments.tests.factories.legacy import (  # noqa: F401
+    ExperimentBucketNamespaceFactory,
+    ExperimentBucketRangeFactory,
+    ExperimentChangeLogFactory,
+    ExperimentCommentFactory,
+    ExperimentControlFactory,
+    ExperimentFactory,
+    ExperimentRapidFactory,
+    ExperimentVariantFactory,
+    VariantPreferencesFactory,
+)
+from experimenter.experiments.tests.factories.nimbus import (  # noqa: F401
+    NimbusExperimentFactory,
+)
