+++ app/experimenter/experiments/tests/test_changelog_utils.py
@@ -9,9 +9,9 @@ from experimenter.experiments.models import Experiment
 from experimenter.experiments.tests.factories import (
     ExperimentFactory,
     ExperimentVariantFactory,
-    UserFactory,
     VariantPreferencesFactory,
 )
+from experimenter.openidc.tests.factories import UserFactory
 from experimenter.projects.tests.factories import ProjectFactory
